+++ test/thump_test.dart
@@ -89,8 +89,8 @@ void main() {
     world.add(block, AABB.xywh(x: 64, y: 16, width: 16, height: 16));
     world.add(man, AABB.xywh(x: 47, y: 16, width: 16, height: 16));
     MoveResult result = world.move(man, 2, 0);
-    expect(result.x, closeTo(48.0, 0.01));
-    expect(result.y, closeTo(16.0, 0.01));
+    expect(result.position.x, closeTo(48.0, 0.01));
+    expect(result.position.y, closeTo(16.0, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
     expect(result.collisions[0].edge, Edge.right);
@@ -103,8 +103,8 @@ void main() {
     world.add(block, AABB.xywh(x: 64, y: 16, width: 1, height: 1));
     world.add(man, AABB.xywh(x: 61.99, y: 16, width: 1, height: 1));
     MoveResult result = world.move(man, 2.99, 0);
-    expect(result.x, closeTo(63.0, 0.01));
-    expect(result.y, closeTo(16.0, 0.01));
+    expect(result.position.x, closeTo(63.0, 0.01));
+    expect(result.position.y, closeTo(16.0, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
   });
@@ -131,8 +131,8 @@ void main() {
     world.add(block, AABB.xywh(x: 48, y: 32, width: 16, height: 16));
     world.add(man, AABB.xywh(x: 48, y: 16, width: 16, height: 16));
     MoveResult result = world.move(man, 3, 4);
-    expect(result.x, closeTo(51.0, 0.01));
-    expect(result.y, closeTo(16.0, 0.01));
+    expect(result.position.x, closeTo(51.0, 0.01));
+    expect(result.position.y, closeTo(16.0, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
     expect(result.collisions[0].edge, Edge.bottom);
@@ -145,8 +145,8 @@ void main() {
     world.add(block, AABB.xywh(x: 48, y: 32, width: 16, height: 16));
     world.add(man, AABB.xywh(x: 48, y: 16, width: 16, height: 16));
     MoveResult result = world.move(man, 3, 3);
-    expect(result.x, closeTo(51, 0.01));
-    expect(result.y, closeTo(16.0, 0.01));
+    expect(result.position.x, closeTo(51, 0.01));
+    expect(result.position.y, closeTo(16.0, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
   });
@@ -158,8 +158,8 @@ void main() {
     world.add(block, AABB.xywh(x: 64, y: 16, width: 16, height: 16));
     world.add(man, AABB.xywh(x: 47, y: 16, width: 16, height: 16));
     MoveResult result = world.move(man, 2, 0, handler: (obj) => Behavior.Pass);
-    expect(result.x, closeTo(49.0, 0.01));
-    expect(result.y, closeTo(16.0, 0.01));
+    expect(result.position.x, closeTo(49.0, 0.01));
+    expect(result.position.y, closeTo(16.0, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
   });
@@ -171,8 +171,8 @@ void main() {
     world.add(block, AABB.xywh(x: 64, y: 16, width: 16, height: 16));
     world.add(man, AABB.xywh(x: 81, y: 16, width: 16, height: 16));
     MoveResult result = world.move(man, -2, 0);
-    expect(result.x, closeTo(80.0, 0.01));
-    expect(result.y, closeTo(16.0, 0.01));
+    expect(result.position.x, closeTo(80.0, 0.01));
+    expect(result.position.y, closeTo(16.0, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
     expect(result.collisions[0].edge, Edge.left);
@@ -185,8 +185,8 @@ void main() {
     world.add(block, AABB.xywh(x: 64, y: 16, width: 16, height: 16));
     world.add(man, AABB.xywh(x: 64, y: 33, width: 16, height: 16));
     MoveResult result = world.move(man, 0, -2);
-    expect(result.x, closeTo(64.0, 0.01));
-    expect(result.y, closeTo(32.0, 0.01));
+    expect(result.position.x, closeTo(64.0, 0.01));
+    expect(result.position.y, closeTo(32.0, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
     expect(result.collisions[0].edge, Edge.top);
@@ -200,8 +200,8 @@ void main() {
     world.add(block, AABB.xywh(x: 16, y: 32, width: 16, height: 16));
     world.add(man, AABB.xywh(x: 0, y: 50, width: 16, height: 16));
     MoveResult result = world.move(man, 0, -50);
-    expect(result.x, closeTo(0.0, 0.01));
-    expect(result.y, closeTo(0.0, 0.01));
+    expect(result.position.x, closeTo(0.0, 0.01));
+    expect(result.position.y, closeTo(0.0, 0.01));
     expect(result.collisions.length, 0);
   });
 
@@ -212,8 +212,8 @@ void main() {
     world.add(block, AABB.xywh(x: 64, y: 32, width: 16, height: 16));
     world.add(man, AABB.xywh(x: 64, y: 15, width: 16, height: 16));
     MoveResult result = world.move(man, 0, 2);
-    expect(result.x, closeTo(64.0, 0.01));
-    expect(result.y, closeTo(16.0, 0.01));
+    expect(result.position.x, closeTo(64.0, 0.01));
+    expect(result.position.y, closeTo(16.0, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
 
@@ -235,8 +235,8 @@ void main() {
     world.add(man, AABB.xywh(x: 47.9, y: 16, width: 16, height: 16));
     MoveResult result =
         world.move(man, 10, 10, handler: (obj) => Behavior.Touch);
-    expect(result.x, closeTo(48.0, 0.01));
-    expect(result.y, closeTo(16.1, 0.01));
+    expect(result.position.x, closeTo(48.0, 0.01));
+    expect(result.position.y, closeTo(16.1, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
   });
@@ -249,8 +249,8 @@ void main() {
     world.add(man, AABB.xywh(x: 80.1, y: 16, width: 16, height: 16));
     MoveResult result =
         world.move(man, -10, 10, handler: (obj) => Behavior.Touch);
-    expect(result.x, closeTo(80.0, 0.01));
-    expect(result.y, closeTo(16.1, 0.01));
+    expect(result.position.x, closeTo(80.0, 0.01));
+    expect(result.position.y, closeTo(16.1, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
   });
@@ -263,8 +263,8 @@ void main() {
     world.add(man, AABB.xywh(x: 64, y: 32.1, width: 16, height: 16));
     MoveResult result =
         world.move(man, 10, -10, handler: (obj) => Behavior.Touch);
-    expect(result.y, closeTo(32.0, 0.01));
-    expect(result.x, closeTo(64.1, 0.01));
+    expect(result.position.y, closeTo(32.0, 0.01));
+    expect(result.position.x, closeTo(64.1, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
   });
@@ -277,8 +277,8 @@ void main() {
     world.add(man, AABB.xywh(x: 64, y: 15.9, width: 16, height: 16));
     MoveResult result =
         world.move(man, -10, 10, handler: (obj) => Behavior.Touch);
-    expect(result.y, closeTo(16.0, 0.01));
-    expect(result.x, closeTo(63.9, 0.01));
+    expect(result.position.y, closeTo(16.0, 0.01));
+    expect(result.position.x, closeTo(63.9, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
   });
@@ -292,8 +292,8 @@ void main() {
     world.add(blockDown, AABB.xywh(x: 32, y: 48, width: 16, height: 16));
     world.add(man, AABB.xywh(x: 31, y: 31, width: 16, height: 16));
     MoveResult result = world.move(man, 2, 2);
-    expect(result.y, closeTo(32.0, 0.01));
-    expect(result.x, closeTo(32.0, 0.01));
+    expect(result.position.y, closeTo(32.0, 0.01));
+    expect(result.position.x, closeTo(32.0, 0.01));
     expect(result.collisions.length, 2);
   });
 
@@ -305,10 +305,10 @@ void main() {
     world.add(man, AABB.xywh(x: 48, y: 14, width: 16, height: 16));
     MoveResult result =
         world.move(man, 4, 4, handler: (obj) => Behavior.Bounce);
-    expect(result.x, closeTo(52.0, 0.01));
-    expect(result.y, closeTo(14.0, 0.01));
-    expect(result.dx, closeTo(sqrt2Reciprocal, 0.01));
-    expect(result.dy, closeTo(-sqrt2Reciprocal, 0.01));
+    expect(result.position.x, closeTo(52.0, 0.01));
+    expect(result.position.y, closeTo(14.0, 0.01));
+    expect(result.direction.x, closeTo(sqrt2Reciprocal, 0.01));
+    expect(result.direction.y, closeTo(-sqrt2Reciprocal, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
   });
@@ -321,10 +321,10 @@ void main() {
     world.add(man, AABB.xywh(x: 48, y: 50, width: 16, height: 16));
     MoveResult result =
         world.move(man, 4, -4, handler: (obj) => Behavior.Bounce);
-    expect(result.x, closeTo(52.0, 0.01));
-    expect(result.y, closeTo(50.0, 0.01));
-    expect(result.dx, closeTo(sqrt2Reciprocal, 0.01));
-    expect(result.dy, closeTo(sqrt2Reciprocal, 0.01));
+    expect(result.position.x, closeTo(52.0, 0.01));
+    expect(result.position.y, closeTo(50.0, 0.01));
+    expect(result.direction.x, closeTo(sqrt2Reciprocal, 0.01));
+    expect(result.direction.y, closeTo(sqrt2Reciprocal, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
   });
@@ -337,10 +337,10 @@ void main() {
     world.add(man, AABB.xywh(x: 30, y: 32, width: 16, height: 16));
     MoveResult result =
         world.move(man, 4, 4, handler: (obj) => Behavior.Bounce);
-    expect(result.x, closeTo(30.0, 0.01));
-    expect(result.y, closeTo(36.0, 0.01));
-    expect(result.dx, closeTo(-sqrt2Reciprocal, 0.01));
-    expect(result.dy, closeTo(sqrt2Reciprocal, 0.01));
+    expect(result.position.x, closeTo(30.0, 0.01));
+    expect(result.position.y, closeTo(36.0, 0.01));
+    expect(result.direction.x, closeTo(-sqrt2Reciprocal, 0.01));
+    expect(result.direction.y, closeTo(sqrt2Reciprocal, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
   });
@@ -353,10 +353,10 @@ void main() {
     world.add(man, AABB.xywh(x: 66, y: 32, width: 16, height: 16));
     MoveResult result =
         world.move(man, -4, 4, handler: (obj) => Behavior.Bounce);
-    expect(result.x, closeTo(66.0, 0.01));
-    expect(result.y, closeTo(36.0, 0.01));
-    expect(result.dx, closeTo(sqrt2Reciprocal, 0.01));
-    expect(result.dy, closeTo(sqrt2Reciprocal, 0.01));
+    expect(result.position.x, closeTo(66.0, 0.01));
+    expect(result.position.y, closeTo(36.0, 0.01));
+    expect(result.direction.x, closeTo(sqrt2Reciprocal, 0.01));
+    expect(result.direction.y, closeTo(sqrt2Reciprocal, 0.01));
     expect(result.collisions.length, 1);
     expect(result.collisions[0].object, block);
     expect(result.collisions[0].behavior, Behavior.Bounce);