+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added a Collision object and added the collided edge to the MoveResults.
 - Started returning the dx and dy of the resulting object.
 - Started reporting the behavior that was used for collisions.
+- Introduced Vector2.
 
 ## 0.1.1